fix LP64 real (double) constant load/store: 32-bit union halves
Real *constants* in the .dis data section loaded as ~0 on LP64 (so all
floating-point literals, and any math through them, were wrong), while reals
computed at run time were fine. Two functions reassembled/split an IEEE double
through a `union { double d; unsigned long ul[2]; }`: on LP64 `unsigned
long`/`ulong` is 8 bytes, so ul[0] aliased the whole double and ul[1] was
garbage, emitting/reading the wrong two 32-bit words. Pin the union element to
`unsigned int` (32-bit on both ILP32 and LP64), the same fix already applied to
libmath/dtoa.c in stage 1.

  - limbo/dtocanon.c: dtocanon() (compiler: splits a double into the two words
    written to the data section) and its canontod().
  - libinterp/load.c: canontod() (loader: reassembles the double from the data
    section).

dtocanon is in the compiler, so the dis tree must be recompiled (done in the
following commit). The self-host compiler (appl/cmd/limbo) is unaffected: it
serialises reals via the Math export_real builtin, not a C union.

Verified on aarch64/LP64: real arithmetic, sqrt/sin/cos/atan/pow/log/exp,
int<->real conversion, 1e+-300 extremes (dtoa path), real arrays, and
string->real parsing all produce correct results.

+++ limbo/dtocanon.c
@@ -3,7 +3,14 @@
 void
 dtocanon(double f, ulong v[])
 {
-	union { double d; ulong ul[2]; } a;
+	/*
+	 * Split an IEEE double into its two 32-bit halves for the .dis data
+	 * section.  The union element MUST be 32-bit: on LP64 "ulong" is 8 bytes,
+	 * so ul[0] would alias the whole double and ul[1] garbage, emitting the
+	 * wrong words (every real constant then loaded as ~0).  "unsigned int" is
+	 * 32-bit on both ILP32 and LP64 (cf. canontod, libmath/dtoa.c).
+	 */
+	union { double d; unsigned int ul[2]; } a;
 
 	a.d = 1.;
 	if(a.ul[0]){
@@ -20,7 +27,7 @@ dtocanon(double f, ulong v[])
 double
 canontod(ulong v[2])
 {
-	union { double d; unsigned long ul[2]; } a;
+	union { double d; unsigned int ul[2]; } a;	/* 32-bit halves; see dtocanon */
 
 	a.d = 1.;
 	if(a.ul[0]) {
